fix(rust): Add bounds checks to prevent panic on malformed CEA-708 data
Fixes #1616 - Segmentation fault when extracting from MP4 remuxed from HLS

The CEA-708 decoder could panic when processing truncated or malformed
caption data blocks:

1. Fixed EXT1 command handling in process_service_block():
   - Changed &block[1..] to &block[(i+1)..] for correct slice offset
   - Added bounds check before accessing the next byte after EXT1

2. Added bounds checks in handle_extended_char():
   - Check for empty block before accessing block[0]
   - Check block.len() >= 2 before accessing block[1] for C3 commands

3. Removed unnecessary `as i64` cast in es/pic.rs to fix clippy warning

Added 4 unit tests to verify the bounds checking:
- test_handle_extended_char_empty_block
- test_handle_extended_char_c3_insufficient_bytes
- test_process_service_block_ext1_at_end
- test_process_service_block_ext1_with_truncated_c3

+++ src/rust/src/es/pic.rs
@@ -278,7 +278,7 @@ pub unsafe fn read_pic_info(
     if ccx_options.use_gop_as_pts == 1 {
         // Calculate current FTS based on GOP start time + frame offset
         let frame_offset_ms = (dec_ctx.frames_since_last_gop as f64 * 1000.0 / current_fps) as i64;
-        (*dec_ctx.timing).fts_now = fts_at_gop_start as i64 + frame_offset_ms;
+        (*dec_ctx.timing).fts_now = fts_at_gop_start + frame_offset_ms;
 
         // Update fts_max if needed
         if (*dec_ctx.timing).fts_now > (*dec_ctx.timing).fts_max {
